requirements.txt
matplotlib
yfinance
tkinter
PIL
pathlib
functools
pandas_datareader.data
numpy
pandas
time
